make tests python3 compatible
